requirements.txt
flask
flask_cors
pillow
opencv-python